change double to int to allow reading

+++ codebase/src/main/java/ethz/ivt/externalities/unchosenAlternatives/AlternativeRecord.java
@@ -59,7 +59,7 @@ public class AlternativeRecord {
 	Double frequency;
 	
 	@CsvBindByName
-	Integer transfers;
+	Double transfers;
 	
 	@CsvBindByName
 	private boolean chosen = false;
